fix: add minor review changes
